# openapi.yaml
openapi: 3.0.0
info:
  title: FinTrack API
  description: |
    FinTrack adalah aplikasi fintech berbasis Laravel 12 dengan arsitektur Modular Monolith.
    API ini menyediakan tiga modul utama:
    - **Transaction Management**: Pencatatan dan kategorisasi transaksi dengan AI
    - **Financial Analysis**: Analisis finansial dengan AI-powered insights
    - **Financial Planning**: Goal-based financial planning

    ## Fitur Utama
    - 🔐 Autentikasi Token-based dengan Laravel Sanctum
    - 🤖 AI-powered transaction categorization
    - 📊 Real-time financial analytics
    - 🎯 Goal tracking dan progress monitoring

    ## Environment Setup
    - Base URL: `http://localhost:8000/api`
    - Autentikasi: Gunakan Bearer token yang diterima dari endpoint login
    - Format request body: `application/json`

  version: 1.0.0
  contact:
    name: FinTrack Support
    email: support@fintrack.dev
    url: https://github.com/yourusername/uas
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.fintrack.dev/api
    description: Production server
...
paths:
...
  /transactions:
    get:
      tags:
        - Transactions
      summary: List Semua Transaksi User
      description: |
        Retrieve semua transaksi milik user yang login.
        - Diurutkan berdasarkan date descending (terbaru dulu)
        - Include detail amount, type, category, description
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List transaksi berhasil diambil
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
              example:
                - id: 1
                  user_id: 1
                  amount: "50000.00"
                  type: expense
                  category: Food
                  description: Lunch at cafe
                  date: 2026-07-01
                  created_at: 2026-07-01T10:30:00Z
                  updated_at: 2026-07-01T10:30:00Z
                - id: 2
                  user_id: 1
                  amount: "150000.00"
                  type: income
                  category: Salary
                  description: Monthly salary
                  date: 2026-07-01
                  created_at: 2026-07-01T08:00:00Z
                  updated_at: 2026-07-01T08:00:00Z

        '401':
          description: Unauthorized - Token tidak valid/expired

    post:
      tags:
        - Transactions
      summary: Buat Transaksi Baru
      description: |
        Membuat transaksi baru untuk user yang login.
        - Kategori bisa di-autofill oleh AI dari description
        - Atau manual input kategori
        - Amount dalam decimal, type: income atau expense
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - type
                - description
                - date
              properties:
                amount:
                  type: number
                  format: decimal
                  example: 50000
                  description: Jumlah transaksi
                type:
                  type: string
                  enum: [income, expense]
                  example: expense
                  description: Tipe transaksi
                category:
                  type: string
                  example: Food
                  description: Kategori transaksi (opsional, auto-filled by AI)
                description:
                  type: string
                  example: Lunch at cafe
                  description: Deskripsi transaksi
                date:
                  type: string
                  format: date
                  example: 2026-07-01
                  description: Tanggal transaksi
      responses:
        '201':
          description: Transaksi berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
              example:
                id: 3
                user_id: 1
                amount: "50000.00"
                type: expense
                category: Food
                description: Lunch at cafe
                date: 2026-07-01
                created_at: 2026-07-01T12:00:00Z
                updated_at: 2026-07-01T12:00:00Z

        '422':
          description: Validasi gagal
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        amount:
          type: string
          format: decimal
          example: "50000.00"
        type:
          type: string
          enum: [income, expense]
          example: expense
        category:
          type: string
          example: Food
        description:
          type: string
          example: Lunch at cafe
        date:
          type: string
          format: date
          example: 2026-07-01
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: |
        Gunakan token yang diterima dari endpoint /login atau /register.
        Tambahkan ke header: `Authorization: Bearer {token}`